--- xo-starter-ienum/src/test/java/com/cntest/su/ienum/jackson/TestBean.java
@@ -1,13 +1,0 @@
-package com.cntest.su.ienum.jackson;
-
-public class TestBean {
-  private TestEnum num = TestEnum.ONE;
-
-  public TestEnum getNum() {
-    return num;
-  }
-
-  public void setNum(TestEnum num) {
-    this.num = num;
-  }
-}
