+++ xo-starter-ienum/src/test/java/com/github/jnoee/xo/ienum/TestBean.java
@@ -1,0 +1,13 @@
+package com.github.jnoee.xo.ienum;
+
+public class TestBean {
+  private TestEnum num = TestEnum.ONE;
+
+  public TestEnum getNum() {
+    return num;
+  }
+
+  public void setNum(TestEnum num) {
+    this.num = num;
+  }
+}
